refactor: remove unwanted code
