+++ src/domain/entities/DataForm.ts
@@ -1,26 +1,10 @@
-import i18n from "../../locales";
+import { DataElementType } from "../../data/repositories/bulk-load/EGASPProgramDefaultRepository";
 import { Id, NamedRef } from "./Ref";
 
-export const dataFormTypes = ["dataSets", "programs", "trackerPrograms"] as const;
-export type DataFormType = typeof dataFormTypes[number];
 export type DataFormPeriod = "Daily" | "Monthly" | "Yearly" | "Weekly";
 
-export function ofType<T>(obj: T): T {
-    return obj;
-}
-
-export function getTranslations() {
-    return {
-        dataFormTypes: ofType<Record<DataFormType, string>>({
-            dataSets: i18n.t("Data Set"),
-            programs: i18n.t("Event Program"),
-            trackerPrograms: i18n.t("Tracker Program"),
-        }),
-    };
-}
-
 export interface DataForm {
-    type: DataFormType;
+    type: "programs";
     id: Id;
     name: string;
     periodType?: DataFormPeriod;
@@ -54,37 +38,4 @@ export interface DataElement {
     valueType: DataElementType;
     categoryOptionCombos?: Array<{ id: Id; name: string }>;
     options: Array<{ id: Id; code: string }>;
-}
-
-export type DataElementType =
-    | "TEXT"
-    | "LONG_TEXT"
-    | "LETTER"
-    | "PHONE_NUMBER"
-    | "EMAIL"
-    | "BOOLEAN"
-    | "TRUE_ONLY"
-    | "DATE"
-    | "DATETIME"
-    | "TIME"
-    | "NUMBER"
-    | "UNIT_INTERVAL"
-    | "PERCENTAGE"
-    | "INTEGER"
-    | "INTEGER_POSITIVE"
-    | "INTEGER_NEGATIVE"
-    | "INTEGER_ZERO_OR_POSITIVE"
-    | "TRACKER_ASSOCIATE"
-    | "USERNAME"
-    | "COORDINATE"
-    | "ORGANISATION_UNIT"
-    | "AGE"
-    | "URL"
-    | "FILE_RESOURCE"
-    | "IMAGE";
-
-export interface OrganisationUnit {
-    id: Id;
-    name: string;
-    path: string;
 }
